third_party/blink/web_tests/fast/events/touch/gesture/gesture-tap-cancel-hover-state-expected.txt

Gestures go here

Tests that gesture tapcancel clears hover properly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

See http://crbug.com/316974 for details

PASS successfullyParsed is true

TEST COMPLETE

Verify hover, active aren't initially set
PASS getHoverActiveState(box) is "default"
tapCancel on element when it is Active should keep hover
PASS getHoverActiveState(box) is "hoveredAndActive"
PASS getHoverActiveState(box) is "hovered"
tapCancel outside element when it is Active should clear it
PASS getHoverActiveState(box) is "hoveredAndActive"
PASS getHoverActiveState(box) is "default"
tapCancel on element when it is hovered but not active should keep hover
PASS getHoverActiveState(box) is "hovered"
PASS getHoverActiveState(box) is "hovered"
tapCancel outside element when it is hovered but not active should keep hover
PASS getHoverActiveState(box) is "hovered"
PASS getHoverActiveState(box) is "hovered"
